move all holochain and holo code into its own store
