+++ rcnotes.txt
@@ -77,8 +77,7 @@ signal:
 TODO
   store current channel id in localstorage
 
-
 Questions
   Is there a better way to separate out the two clients / modes (HC and Holo), such that the rest of the app is insulated from it?
   Is there a better way to handle the initialization? Calling initializeApp from initializeAgent feels weird to me.
-
+  Should channel be stored in addition to channels? Alternative is just to store "currentChannelId", and create a getter to fetch the actual channel
